refine protocol, use int64_t timestamp for ts and jitter.

+++ trunk/src/core/srs_core_protocol.hpp
@@ -204,8 +204,9 @@ struct SrsMessageHeader
 	* Four-byte field that contains a timestamp of the message.
 	* The 4 bytes are packed in the big-endian order.
 	* @remark, used as calc timestamp when decode and encode time.
+	* @remark, we use 64bits for large time for jitter detect and hls.
 	*/
-	u_int32_t timestamp;
+	int64_t timestamp;
 	
 	SrsMessageHeader();
 	virtual ~SrsMessageHeader();
@@ -1126,7 +1127,7 @@ int srs_rtmp_expect_message(SrsProtocol* protocol, SrsCommonMessage** pmsg, T** 
 		T* pkt = dynamic_cast<T*>(msg->get_packet());
 		if (!pkt) {
 			delete msg;
-			srs_trace("drop message(type=%d, size=%d, time=%d, sid=%d).", 
+			srs_trace("drop message(type=%d, size=%d, time=%"PRId64", sid=%d).", 
 				msg->header.message_type, msg->header.payload_length,
 				msg->header.timestamp, msg->header.stream_id);
 			continue;
